create button only works when endpoint exists
