+++ src/app/api/instances/[name]/alarms/slack/route.ts
@@ -5,7 +5,9 @@ export async function POST(request: NextRequest) {
   const body = await request.json();
   const webhookUrl = body.webhookUrl;
 
-  if (!webhookUrl)
+  console.log("webhookUrl", webhookUrl);
+
+  if (!("webhookUrl" in body))
     return new NextResponse("Must have a webhookUrl in the request body.", {
       status: 400,
     });
